requirements.txt
black==18.9b0
inflection==0.3.1
mypy==0.630
pytest==3.8
requests==2.18.4
singer-python==5.2.3
snowflake-connector-python==1.6.10
snowflake-sqlalchemy>=1.1.2
